refactor(host-application): split task workflow services
