+++ apps/desktop/src-tauri/crates/host-application/src/app_service/mod.rs
@@ -65,9 +65,9 @@ pub(crate) use startup_metrics::{
 pub(crate) use workflow_rules::{
     can_replace_epic_subtask_status, can_set_plan, can_set_spec_from_status,
     default_qa_required_for_issue_type, derive_agent_workflows, derive_available_actions,
-    normalize_required_markdown, normalize_subtask_plan_inputs, normalize_title_key,
-    validate_parent_relationships_for_create, validate_parent_relationships_for_update,
-    validate_plan_subtask_rules, validate_transition,
+    is_open_state, normalize_required_markdown, normalize_subtask_plan_inputs,
+    normalize_title_key, validate_parent_relationships_for_create,
+    validate_parent_relationships_for_update, validate_plan_subtask_rules, validate_transition,
 };
 
 #[cfg(test)]
