+++ apps/desktop/src-tauri/crates/host-application/src/app_service/task_workflow/mod.rs
@@ -1,0 +1,16 @@
+pub(super) use super::{
+    can_replace_epic_subtask_status, can_set_plan, can_set_spec_from_status,
+    default_qa_required_for_issue_type, is_open_state, normalize_required_markdown,
+    normalize_subtask_plan_inputs, normalize_title_key,
+    validate_parent_relationships_for_create, validate_parent_relationships_for_update,
+    validate_plan_subtask_rules, validate_transition, AppService,
+};
+
+mod document_service;
+mod qa_service;
+mod session_service;
+mod task_context;
+mod task_service;
+
+#[cfg(test)]
+mod tests;
